Add long form sprite node constructors
Up to now, we accidentally used a short form for all sprite node ocnstructors and a longer form (acutally matching the Objective-C name) for all others.

TODO: We need to either deprecate the short form version for sprite nodes or add long form version for all other types of nodes.

Contributed to #5